packages/platform/backend/pyproject.toml
[project]
name = "lingnexus-platform"
version = "1.0.0"
description = "Low-code platform for building AI agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic[email]>=2.5.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # 依赖本地 framework（通过工作区）
    "lingnexus-framework",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# 指定包的位置
[tool.hatch.build.targets.wheel]
packages = ["db", "api", "core", "models", "services"]

[tool.uv.sources]
lingnexus-framework = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
